Updates to refactor automation project
