+++ automation/build.gradle
@@ -1,0 +1,19 @@
+apply plugin: 'java-library'
+apply plugin: 'net.serenity-bdd.aggregator'
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation 'net.serenity-bdd:serenity-core:1.9.6'
+    testImplementation 'net.serenity-bdd:serenity-junit:1.9.6'
+    testImplementation 'net.serenity-bdd:serenity-screenplay:1.9.6'
+    testImplementation 'net.serenity-bdd:serenity-screenplay-webdriver:1.9.6'
+    testImplementation 'org.slf4j:slf4j-simple:1.7.7'
+    testImplementation 'org.assertj:assertj-core:1.7.0'
+    testImplementation 'junit:junit:4.12'
+    testImplementation 'io.appium:java-client:5.0.4'
+}
+
+gradle.startParameter.continueOnFailure = true
+
+sourceCompatibility = "1.7"
+targetCompatibility = "1.7"
