+++ automation/src/main/java/com/example/automation/Automation.java
@@ -1,0 +1,4 @@
+package com.example.automation;
+
+public class Automation {
+}
