+++ automation/src/test/java/com/microsoft/identity/common/test/automation/actors/User.java
@@ -1,4 +1,4 @@
-package com.microsoft.aad.automation.testapp.actors;
+package com.microsoft.identity.common.test.automation.actors;
 
 import net.serenitybdd.screenplay.Actor;
