+++ automation/src/test/java/com/microsoft/identity/common/test/automation/interactions/PressHome.java
@@ -1,0 +1,19 @@
+package com.microsoft.identity.common.test.automation.interactions;
+
+import net.serenitybdd.screenplay.Actor;
+import net.serenitybdd.screenplay.Interaction;
+import net.serenitybdd.screenplay.abilities.BrowseTheWeb;
+import net.thucydides.core.webdriver.WebDriverFacade;
+
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.android.AndroidKeyCode;
+
+public class PressHome implements Interaction {
+
+    @Override
+    public <T extends Actor> void performAs(T actor) {
+        WebDriverFacade facade = (WebDriverFacade)BrowseTheWeb.as(actor).getDriver();
+        AndroidDriver androidDriver = (AndroidDriver)facade.getProxiedDriver();
+        androidDriver.pressKeyCode(AndroidKeyCode.HOME);
+    }
+}
