+++ serenity.properties
@@ -1,0 +1,38 @@
+# Define the default driver
+#webdriver.driver=phantomjs
+
+# Appears at the top of the reports
+serenity.project.name = Automation for ADAL, MSAL using Appium, Serenity Screenplay
+
+# Root package for any JUnit acceptance tests
+#serenity.test.root=net.serenitybdd.tutorials.features
+
+# Customise your riequirements hierarchy
+#serenity.requirement.types=feature, story
+
+# Run the tests without calling webdriver - useful to check your JUnit wiring
+#serenity.dry.run=true
+
+# Customise browser size
+#serenity.browser.height = 1200
+#serenity.browser.width = 1200
+
+webdriver.driver= appium
+#webdriver.timeouts.implicitlywait = 50000
+#webdriver.wait.for.timeout = 100000
+
+
+#appium.hub = http://localhost:4723/wd/hub
+####### iOS CAPS ######
+#appium.automationName = XCUITest
+#appium.platformName=IOS
+#appium.platformVersion = 10.3
+#appium.deviceName  = iPhone 7
+#appium.app =  /Users/vikramanna/Documents/vikram/AppiumSerenityPOC/WordPress.app
+####### Android CAPS ######
+appium.automationName=uiautomator2
+appium.platformName=Android
+appium.platformVersion=7.1.2
+appium.deviceName=Pixel XL
+appium.app=C:\\temp\\android-unified\\automationtestapp\\outputs\\apk\\adal\\debug\\automationtestapp-adal-debug.apk
+appium.newCommandTimeout=1000
