+++ settings.gradle
@@ -1,3 +1,3 @@
-include ':adal', ':userappwithbroker', ':keyvault', ':labapi', ":automationtestapp", ':common'
+include ':adal', ':userappwithbroker', ':keyvault', ':labapi', ':automation', ":automationtestapp", ':common'
 project(':common').projectDir = new File('common/common')
